Move the ClientRegistry type to its own module
And adjust imports & exports.
