+++ ouroboros-network/src/Ouroboros/Network/BlockFetch/ClientRegistry.hs
@@ -1,0 +1,97 @@
+{-# LANGUAGE NamedFieldPuns             #-}
+{-# LANGUAGE RecordWildCards            #-}
+{-# LANGUAGE FlexibleContexts           #-}
+{-# LANGUAGE BangPatterns               #-}
+
+module Ouroboros.Network.BlockFetch.ClientRegistry (
+    -- * Registry of block fetch clients
+    FetchClientRegistry(..),
+    newFetchClientRegistry,
+    bracketFetchClient,
+    readFetchClientsStatus,
+    readFetchClientsStateVars,
+    readFetchClientsStates,
+    readFetchClientsReqVars,
+  ) where
+
+import qualified Data.Map as Map
+import           Data.Map (Map)
+
+import           Control.Monad.Class.MonadSTM
+import           Control.Monad.Class.MonadThrow
+
+import           Ouroboros.Network.BlockFetch.ClientState
+
+
+
+-- | A registry for the threads that are executing the client side of the
+-- 'BlockFetch' protocol to communicate with our peers.
+--
+-- The registry contains the shared variables we use to communicate with these
+-- threads, both to track their status and to provide instructions.
+--
+-- The threads add\/remove themselves to\/from this registry when they start up
+-- and shut down.
+--
+newtype FetchClientRegistry peer header m =
+        FetchClientRegistry (TVar m (Map peer (FetchClientStateVars header m)))
+
+newFetchClientRegistry :: MonadSTM m => m (FetchClientRegistry peer header m)
+newFetchClientRegistry = FetchClientRegistry <$> newTVarM Map.empty
+
+bracketFetchClient :: (MonadThrow m, MonadSTM m, Ord peer)
+                   => FetchClientRegistry peer header m
+                   -> peer
+                   -> (FetchClientStateVars header m -> m a)
+                   -> m a
+bracketFetchClient (FetchClientRegistry registry) peer =
+    bracket register unregister
+  where
+    register = atomically $ do
+      stateVars <- newFetchClientStateVars
+      modifyTVar' registry (Map.insert peer stateVars)
+      return stateVars
+
+    unregister FetchClientStateVars{fetchClientStatusVar} =
+      atomically $ do
+        writeTVar fetchClientStatusVar PeerFetchStatusShutdown
+        modifyTVar' registry (Map.delete peer)
+
+-- | A read-only 'STM' action to get the current 'PeerFetchStatus' for all
+-- fetch clients in the 'FetchClientRegistry'.
+--
+readFetchClientsStatus :: MonadSTM m
+                       => FetchClientRegistry peer header m
+                       -> STM m (Map peer (PeerFetchStatus header))
+readFetchClientsStatus (FetchClientRegistry registry) =
+  readTVar registry >>= traverse (readTVar . fetchClientStatusVar)
+
+-- | A read-only 'STM' action to get the 'FetchClientStateVars' for all fetch
+-- clients in the 'FetchClientRegistry'.
+--
+readFetchClientsStateVars :: MonadSTM m
+                          => FetchClientRegistry peer header m
+                          -> STM m (Map peer (FetchClientStateVars header m))
+readFetchClientsStateVars (FetchClientRegistry registry) = readTVar registry
+
+-- | A read-only 'STM' action to get the current 'PeerFetchStatus' and
+-- 'PeerFetchInFlight' for all fetch clients in the 'FetchClientRegistry'.
+--
+readFetchClientsStates :: MonadSTM m
+                       => FetchClientRegistry peer header m
+                       -> STM m (Map peer (PeerFetchStatus   header,
+                                           PeerFetchInFlight header))
+readFetchClientsStates (FetchClientRegistry registry) =
+  readTVar registry >>=
+  traverse (\s -> (,) <$> readTVar (fetchClientStatusVar s)
+                      <*> readTVar (fetchClientInFlightVar s))
+
+-- | A read-only 'STM' action to get the current 'TFetchRequestVar' for all
+-- fetch clients in the 'FetchClientRegistry'.
+--
+readFetchClientsReqVars :: MonadSTM m
+                        => FetchClientRegistry peer header m
+                        -> STM m (Map peer (TFetchRequestVar m header))
+readFetchClientsReqVars (FetchClientRegistry registry) =
+  readTVar registry >>= return . Map.map fetchClientRequestVar
+
